refactor: changes for alignment to rc.1
